# Use a table lookup for escaping unprintable bytes