@tmthrgd 2018-11-29
#844, #845 and this PR all taken together provide a `-74.90%` time/op delta for `UnpackDomainNameLongestUnprintable`. That is, all three of them together improve the performance of that benchmark by nearly 75%.