@miekg 2018-11-29
> hat is, all three of them together improve the performance of that benchmark by nearly 75%.

holy cow